ai-travel-concierge/requirements.txt
# Core LangChain and LangGraph
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.13
langgraph==0.0.20
langsmith==0.0.77

# LLM Providers
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.5

# Vector Stores and Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1
tiktoken==0.5.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.8
beautifulsoup4==4.12.3

# API Clients
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Data Management
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
tenacity==8.2.3
python-json-logger==2.0.7
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Optional - for specific travel APIs
amadeus==9.0.0
google-maps-services==0.1.0
